Add SQL constraints to impose unicity of (val,val_type) in id2val
id2val contains the properties of ?s ?p ?o triples, so a URI (stored in
val) should be unique together with val_type (0/1/2=uri/bnode/literal,
which is the value’s fundamental type). It was not before this patch:
there were multiple IDs for identical (val, val_type), and as a result
some SPARQL queries resulted in empty results because it was selected
the first ID for a given property, but other results were stored with
other IDs.

With this patch, there is a bijection between (`id`) and (`val`, `val_type`).

A preliminary version of this patch contained also the key
  UNIQUE KEY (id,val(64),val_type),
but it appears this contraint is already filled with both constraints
  PRIMARY KEY (`id`),
  UNIQUE KEY (val(64),val_type),